test: Add e2e test for uploading file

+++ backend/FileService/src/FileService.Application/FilesStorage/IChunkSizeCalculator.cs
@@ -5,5 +5,5 @@ namespace FileService.Application.FilesStorage;
 
 public interface IChunkSizeCalculator
 {
-    Result<(long ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize);
+    Result<(int ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize);
 }
